Now toolbar is being called at home component
