+++ src/App.vue
@@ -1,15 +1,13 @@
 <template>
   <div class="page-container" align="center">
-    <toolbar/>
     <router-view/>
   </div>
 </template>
 
 <script>
-import Toolbar from '@/components/ToolBar.vue';
-export default {
-  components: {
-    Toolbar,
-  },
-};
+import Vue from 'vue';
+import Component from 'vue-class-component';
+
+@Component({})
+export default class App extends Vue {}
 </script>
